modify tenantID type from uint64 to uint32

+++ kv/kv.go
@@ -402,7 +402,7 @@ type Driver interface {
 	// Open returns a new Storage.
 	// The path is the string for storage specific format.
 	Open(path string) (Storage, error)
-	OpenWithTenantID(path string, tenantID uint64) (Storage, error)
+	OpenWithTenantID(path string, tenantID uint32) (Storage, error)
 }
 
 // Storage defines the interface for storage.
